+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+
+* Add `VenmoClient#isReadyToPay()` method
+
 ## 4.1.0
 
 * Bump Cardinal version to `2.2.5-3`